requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv 